call_aggregate

Helper function to run derive_aggregate_spectra.

Usage

call_aggregate(spectra_list, mz_tol, folder_path)

Arguments

spectra_list

A list of spectra objects for all features, each containing the top x percent TIC spectra concatenated.

mz_tol

Mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature.

folder_path

Path to the folder where the data reduction dataframe will be stored.

Value

A list containing:

aggregate_spectra_object: Aggregate spectra object.

aggregate_dataframe: Dataframe of aggregate spectra for all features.

Examples

folder_path <- "path/to/folder"
sp <- preprocess(folder_path, 5, 0.01)
#> *******************Sanity check for input files***********************
#> Error in read_files(folder_path = folder_path, tol_mz, tol_rt): Feature list file not found!
l <- sp[[1]]
#> Error: object 'sp' not found
sp1 <- extract_raw_spectra(folder_path, l, 0.05, 0.8)
#> Error: object 'l' not found
spectra_list <- sp1[[1]]
#> Error: object 'sp1' not found
call_aggregate(spectra_list, 0.05, folder_path)
#> Error: object 'spectra_list' not found