concatenate_spectra

This function extracts MS2 spectra from a list produced by the preprocess function.

Usage

concatenate_spectra(output_list_from_preprocess_function, spectral_files)

Arguments

output_list_from_preprocess_function

A list generated from the preprocess function.

Value

A list containing the MS2 spectra extracted from the mzML files using the predefined tolerances and a vector of features for which MS2 spectra couldn't be extracted.

Details

Reads a list generated from the preprocess function and extracts MS2 spectra within the given mz and RT range. Also returns a list of features for which MS2 spectra couldn't be extracted.

Examples

# Example usage of the function
l1 = preprocess(folder_path = "D:/test_2/", tol_mz = 5, tol_rt = 0.1)  # Replace with actual data
#> *******************Sanity check for input files***********************
#> Error in read_files(folder_path = folder_path, tol_mz, tol_rt): Feature list file not found!
concatenate_spectra(l1)
#> Error: object 'l1' not found