precursor_matching

Matches precursor m/z with library within a given tolerance (As outlined in tol_mz param in preprocess()) and generates a library of reference spectra that can be used for MS/MS-level fragment matching. Also groups fragments in the reference spectra that are close-by within a given tolerance tol

Usage

precursor_matching(l1, folder_path, ionization_mode, tol = 0.01)

Arguments

l1

output from the preprocess()

folder_path

folder containing the required input directory (mzml) and feature list file in .txt format.

ionization_mode

positive or negative. Based on this input, the library data is loaded and stored as a global variable

tol

tolerance to group reference spectra fragments, defaults to 0.01

Value

A dataframe with only those precursors, their mz, RT, etc, which had a corresponding match with the reference library

Examples

# Example usage of the function
precursor_matching(l1, folder_path, "positive", 0.05)
#> Error: object 'folder_path' not found