preprocess

Reads mzml files from the predefined folder path and extracts spectra within the given mz and RT range.

Usage

preprocess(folder_path, tol_mz = 5, tol_rt = 0.1667)

Arguments

folder_path

folder containing the required input directory (mzml) and feature list file in .txt format.

tol_mz

mass tolerance

tol_rt

RT tolerance

Value

A list containing the MS2 spectra extracted from the mzml files using the predefined tolerances and the updated feature list file with MS2 features only

Examples

# Example usage of the function
preprocess(folder_path, 10, 0.1)
#> *******************Sanity check for input files***********************
#> Error: object 'folder_path' not found