tune_parameters

Helper function for generate_denoise_spectra(). Need not be called separately

tune_parameters(agg_df, threshold, name, path)

Arguments

agg_df:

spectrum dataframe with columns Fragment_ID, MZ, Intensity & Freq

threshold:

see function description for generate_denoised_spectra

name:

names of the spectra such as "scan_1.txt" "scan_2.txt" and so on

path:

folder path to store the denoised spectra in .txt and .mzML format

Value

directly prints the spectra to the file in .txt format and returns the spectrum-wise mz and intensity

Examples

# Example usage of the function
tune_parameters(agg_df, threshold, "scan_1.txt", path)
#> Error in eval(expr, envir, enclos): object 'agg_df' not found