tuning module

Determines the optimal frequency cutoff for the selected set of features using a variety of techniques

Usage

tuning_module(folder_path, l4, l5, l6, tolerance = 0.05)

Arguments

folder_path

folder containing the input directory (mzml) feature list files where the results of the matching will be stored.

l4

Contain results of the function label_individual_spectrum.R

l5

Contain the final set of features which matched with the reference at the MS1 level, output of precursor_matching.R

l6

Contain the before denoising matching results

tolerance

fragment ion tolerance at the MS/MS level, defaults to 0.05 Da

Value

A dataframe with matching metrics, annotations and both experimental and reference spectrum identifiers. Features with matching score zero remain unannotated

Examples

# Example usage of the function
fragment_matching_before_denoising(folder_path, l4, tolerance = 0.05, "positive")
#> Error: object 'folder_path' not found