call_aggregate

Helper code to run derive_aggregate_spectra

call_aggregate(spectra_list, mz_tol, folder_path)

Arguments

spectra_list:

list of spectra objects for all features, each containing top x% TIC spectra concatenated

mz_tol:

mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature

folder_path:

the path to the folder where you want to store the data reduction dataframe

Value

a list containing the aggregate spectra object and a dataframe of the aggregate spectra for all features

Examples

# Example usage of the function
folder_path = "/home/shayantan/Desktop/test_package_dures/test_2/"
sp <- preprocess(folder_path, 5, 0.01)
#> *******************Sanity check for input files***********************
#> ✓ Feature list file found!
#> ✓ Feature list file has the correct columns
#> ✓ mzml folder exists
#> RT tolerance file not found. Proceeding with the default mz tolerance of: 5 ppm and RT tolerance of: 0.01 minutes for all samples
#> Reading mzml files
#>   |                                                                              |                                                                      |   0%[1] 1
#>   |                                                                              |                                                                      |   1%[1] 2
#>   |                                                                              |=                                                                     |   1%[1] 3
#>   |                                                                              |=                                                                     |   2%[1] 4
#>   |                                                                              |==                                                                    |   3%[1] 5
#> [1] 6
#>   |                                                                              |===                                                                   |   4%[1] 7
#>   |                                                                              |===                                                                   |   5%[1] 8
#>   |                                                                              |====                                                                  |   5%[1] 9
#>   |                                                                              |====                                                                  |   6%[1] 10
#>   |                                                                              |=====                                                                 |   7%[1] 11
#> [1] 12
#>   |                                                                              |======                                                                |   8%[1] 13
#>   |                                                                              |======                                                                |   9%[1] 14
#>   |                                                                              |=======                                                               |   9%[1] 15
#>   |                                                                              |=======                                                               |  10%[1] 16
#>   |                                                                              |=======                                                               |  11%[1] 17
#>   |                                                                              |========                                                              |  11%[1] 18
#>   |                                                                              |========                                                              |  12%[1] 19
#>   |                                                                              |=========                                                             |  13%[1] 20
#> [1] 21
#>   |                                                                              |==========                                                            |  14%[1] 22
#>   |                                                                              |==========                                                            |  15%[1] 23
#>   |                                                                              |===========                                                           |  15%[1] 24
#>   |                                                                              |===========                                                           |  16%[1] 25
#>   |                                                                              |============                                                          |  17%[1] 26
#> [1] 27
#>   |                                                                              |=============                                                         |  18%[1] 28
#>   |                                                                              |=============                                                         |  19%[1] 29
#>   |                                                                              |==============                                                        |  19%[1] 30
#>   |                                                                              |==============                                                        |  20%[1] 31
#>   |                                                                              |==============                                                        |  21%[1] 32
#>   |                                                                              |===============                                                       |  21%[1] 33
#>   |                                                                              |===============                                                       |  22%[1] 34
#>   |                                                                              |================                                                      |  23%[1] 35
#> [1] 36
#>   |                                                                              |=================                                                     |  24%[1] 37
#>   |                                                                              |=================                                                     |  25%[1] 38
#>   |                                                                              |==================                                                    |  25%[1] 39
#>   |                                                                              |==================                                                    |  26%[1] 40
#>   |                                                                              |===================                                                   |  27%[1] 41
#> [1] 42
#>   |                                                                              |====================                                                  |  28%[1] 43
#>   |                                                                              |====================                                                  |  29%[1] 44
#>   |                                                                              |=====================                                                 |  29%[1] 45
#>   |                                                                              |=====================                                                 |  30%[1] 46
#>   |                                                                              |=====================                                                 |  31%[1] 47
#>   |                                                                              |======================                                                |  31%[1] 48
#>   |                                                                              |======================                                                |  32%[1] 49
#>   |                                                                              |=======================                                               |  33%[1] 50
#> [1] 51
#>   |                                                                              |========================                                              |  34%[1] 52
#>   |                                                                              |========================                                              |  35%[1] 53
#>   |                                                                              |=========================                                             |  35%[1] 54
#>   |                                                                              |=========================                                             |  36%[1] 55
#>   |                                                                              |==========================                                            |  37%[1] 56
#> [1] 57
#>   |                                                                              |===========================                                           |  38%[1] 58
#>   |                                                                              |===========================                                           |  39%[1] 59
#>   |                                                                              |============================                                          |  39%[1] 60
#>   |                                                                              |============================                                          |  40%[1] 61
#>   |                                                                              |============================                                          |  41%[1] 62
#>   |                                                                              |=============================                                         |  41%[1] 63
#>   |                                                                              |=============================                                         |  42%[1] 64
#>   |                                                                              |==============================                                        |  43%[1] 65
#> [1] 66
#>   |                                                                              |===============================                                       |  44%[1] 67
#>   |                                                                              |===============================                                       |  45%[1] 68
#>   |                                                                              |================================                                      |  45%[1] 69
#>   |                                                                              |================================                                      |  46%[1] 70
#>   |                                                                              |=================================                                     |  47%[1] 71
#> [1] 72
#>   |                                                                              |==================================                                    |  48%[1] 73
#>   |                                                                              |==================================                                    |  49%[1] 74
#>   |                                                                              |===================================                                   |  49%[1] 75
#>   |                                                                              |===================================                                   |  50%[1] 76
#>   |                                                                              |===================================                                   |  51%[1] 77
#>   |                                                                              |====================================                                  |  51%[1] 78
#>   |                                                                              |====================================                                  |  52%[1] 79
#>   |                                                                              |=====================================                                 |  53%[1] 80
#> [1] 81
#>   |                                                                              |======================================                                |  54%[1] 82
#>   |                                                                              |======================================                                |  55%[1] 83
#>   |                                                                              |=======================================                               |  55%[1] 84
#>   |                                                                              |=======================================                               |  56%[1] 85
#>   |                                                                              |========================================                              |  57%[1] 86
#> [1] 87
#>   |                                                                              |=========================================                             |  58%[1] 88
#>   |                                                                              |=========================================                             |  59%[1] 89
#>   |                                                                              |==========================================                            |  59%[1] 90
#>   |                                                                              |==========================================                            |  60%[1] 91
#>   |                                                                              |==========================================                            |  61%[1] 92
#>   |                                                                              |===========================================                           |  61%[1] 93
#>   |                                                                              |===========================================                           |  62%[1] 94
#>   |                                                                              |============================================                          |  63%[1] 95
#> [1] 96
#>   |                                                                              |=============================================                         |  64%[1] 97
#>   |                                                                              |=============================================                         |  65%[1] 98
#>   |                                                                              |==============================================                        |  65%[1] 99
#>   |                                                                              |==============================================                        |  66%[1] 100
#>   |                                                                              |===============================================                       |  67%[1] 101
#> [1] 102
#>   |                                                                              |================================================                      |  68%[1] 103
#>   |                                                                              |================================================                      |  69%[1] 104
#>   |                                                                              |=================================================                     |  69%[1] 105
#>   |                                                                              |=================================================                     |  70%[1] 106
#>   |                                                                              |=================================================                     |  71%[1] 107
#>   |                                                                              |==================================================                    |  71%[1] 108
#>   |                                                                              |==================================================                    |  72%[1] 109
#>   |                                                                              |===================================================                   |  73%[1] 110
#> [1] 111
#>   |                                                                              |====================================================                  |  74%[1] 112
#>   |                                                                              |====================================================                  |  75%[1] 113
#>   |                                                                              |=====================================================                 |  75%[1] 114
#>   |                                                                              |=====================================================                 |  76%[1] 115
#>   |                                                                              |======================================================                |  77%[1] 116
#> [1] 117
#>   |                                                                              |=======================================================               |  78%[1] 118
#>   |                                                                              |=======================================================               |  79%[1] 119
#>   |                                                                              |========================================================              |  79%[1] 120
#>   |                                                                              |========================================================              |  80%[1] 121
#>   |                                                                              |========================================================              |  81%[1] 122
#>   |                                                                              |=========================================================             |  81%[1] 123
#>   |                                                                              |=========================================================             |  82%[1] 124
#>   |                                                                              |==========================================================            |  83%[1] 125
#> [1] 126
#>   |                                                                              |===========================================================           |  84%[1] 127
#>   |                                                                              |===========================================================           |  85%[1] 128
#>   |                                                                              |============================================================          |  85%[1] 129
#>   |                                                                              |============================================================          |  86%[1] 130
#>   |                                                                              |=============================================================         |  87%[1] 131
#> [1] 132
#>   |                                                                              |==============================================================        |  88%[1] 133
#>   |                                                                              |==============================================================        |  89%[1] 134
#>   |                                                                              |===============================================================       |  89%[1] 135
#>   |                                                                              |===============================================================       |  90%[1] 136
#>   |                                                                              |===============================================================       |  91%[1] 137
#>   |                                                                              |================================================================      |  91%[1] 138
#>   |                                                                              |================================================================      |  92%[1] 139
#>   |                                                                              |=================================================================     |  93%[1] 140
#> [1] 141
#>   |                                                                              |==================================================================    |  94%[1] 142
#>   |                                                                              |==================================================================    |  95%[1] 143
#>   |                                                                              |===================================================================   |  95%[1] 144
#>   |                                                                              |===================================================================   |  96%[1] 145
#>   |                                                                              |====================================================================  |  97%[1] 146
#> [1] 147
#>   |                                                                              |===================================================================== |  98%[1] 148
#>   |                                                                              |===================================================================== |  99%[1] 149
#>   |                                                                              |======================================================================|  99%[1] 150
#>   |                                                                              |======================================================================| 100%
#>  MS2 data are not present in the following files:
#> [1] "Blank_01,Blank_02,x200708x_Pat_01_POS,x200708x_Pat_02_POS,x200708x_Pat_03_POS,x200708x_Pat_04_POS,x200708x_Pat_05_POS,x200708x_Pat_06_POS,x200708x_Pat_07_POS,x200708x_Pat_08_POS,x200708x_Pat_09_POS,x200708x_Pat_100_POS,x200708x_Pat_101_POS,x200708x_Pat_102_POS,x200708x_Pat_103_POS,x200708x_Pat_104_POS,x200708x_Pat_105_POS,x200708x_Pat_106_POS,x200708x_Pat_107_POS,x200708x_Pat_108_POS,x200708x_Pat_109_POS,x200708x_Pat_10_POS,x200708x_Pat_110_POS,x200708x_Pat_111_POS,x200708x_Pat_112_POS,x200708x_Pat_114_POS,x200708x_Pat_115_POS,x200708x_Pat_116_POS,x200708x_Pat_117_POS,x200708x_Pat_118_POS,x200708x_Pat_119_POS,x200708x_Pat_11_POS,x200708x_Pat_120_POS,x200708x_Pat_121_POS,x200708x_Pat_122_POS,x200708x_Pat_123_POS,x200708x_Pat_124_POS,x200708x_Pat_125_POS,x200708x_Pat_126_POS,x200708x_Pat_127_POS,x200708x_Pat_128_POS,x200708x_Pat_129_POS,x200708x_Pat_12_POS,x200708x_Pat_130_POS,x200708x_Pat_131_POS,x200708x_Pat_132_POS,x200708x_Pat_133_POS,x200708x_Pat_135_POS,x200708x_Pat_137_POS,x200708x_Pat_138_POS,x200708x_Pat_139_POS,x200708x_Pat_13_POS,x200708x_Pat_140_POS,x200708x_Pat_14_POS,x200708x_Pat_15_POS,x200708x_Pat_16_POS,x200708x_Pat_18_POS,x200708x_Pat_19_POS,x200708x_Pat_20_POS,x200708x_Pat_22_POS,x200708x_Pat_23_POS,x200708x_Pat_26_POS,x200708x_Pat_27_POS,x200708x_Pat_28_POS,x200708x_Pat_32_POS,x200708x_Pat_33_POS,x200708x_Pat_34_POS,x200708x_Pat_35_POS,x200708x_Pat_36_POS,x200708x_Pat_37_POS,x200708x_Pat_38_POS,x200708x_Pat_39_POS,x200708x_Pat_40_POS,x200708x_Pat_41_POS,x200708x_Pat_42_POS,x200708x_Pat_43_POS,x200708x_Pat_44_POS,x200708x_Pat_45_POS,x200708x_Pat_46_POS,x200708x_Pat_47_POS,x200708x_Pat_48_POS,x200708x_Pat_49_POS,x200708x_Pat_50_POS,x200708x_Pat_51_POS,x200708x_Pat_52_POS,x200708x_Pat_53_POS,x200708x_Pat_54_POS,x200708x_Pat_55_POS,x200708x_Pat_56_POS,x200708x_Pat_57_POS,x200708x_Pat_58_POS,x200708x_Pat_59_POS,x200708x_Pat_60_POS,x200708x_Pat_61_POS,x200708x_Pat_62_POS,x200708x_Pat_63_POS,x200708x_Pat_64_POS,x200708x_Pat_65_POS,x200708x_Pat_66_POS,x200708x_Pat_67_POS,x200708x_Pat_68_POS,x200708x_Pat_69_POS,x200708x_Pat_70_POS,x200708x_Pat_71_POS,x200708x_Pat_72_POS,x200708x_Pat_73_POS,x200708x_Pat_74_POS,x200708x_Pat_75_POS,x200708x_Pat_76_POS,x200708x_Pat_77_POS,x200708x_Pat_78_POS,x200708x_Pat_79_POS,x200708x_Pat_80_POS,x200708x_Pat_81_POS,x200708x_Pat_82_POS,x200708x_Pat_83_POS,x200708x_Pat_84_POS,x200708x_Pat_85_POS,x200708x_Pat_86_POS,x200708x_Pat_87_POS,x200708x_Pat_89_POS,x200708x_Pat_90_POS,x200708x_Pat_91_POS,x200708x_Pat_92_POS,x200708x_Pat_93_POS,x200708x_Pat_94_POS,x200708x_Pat_95_POS,x200708x_Pat_96_POS,x200708x_Pat_97_POS,x200708x_Pat_98_POS,x200708x_Pat_99_POS,x200708x_QC10_POS,x200708x_QC11_POS,x200708x_QC12_POS,x200708x_QC13_POS,x200708x_QC14_POS,x200708x_QC15_POS,x200708x_QC1_POS,x200708x_QC2_POS,x200708x_QC3_POS,x200708x_QC4_POS,x200708x_QC5_POS,x200708x_QC6_POS,x200708x_QC7_POS,x200708x_QC8_POS,x200708x_QC9_POS"
#> ✓ Spectra has been extracted from the mzml files[1] "Iteration 1"
#> [1] "No MS2 spectra could be extracted for feature 2502 for the given mz and RT range"
#> [1] "Iteration 2"
#> [1] "Iteration 3"
#> [1] "No MS2 spectra could be extracted for feature 4552 for the given mz and RT range"
#> [1] "Iteration 4"
#> [1] "No MS2 spectra could be extracted for feature 3794 for the given mz and RT range"
#> [1] "Iteration 5"
#> [1] "No MS2 spectra could be extracted for feature 3941 for the given mz and RT range"
#> [1] "Iteration 6"
#> [1] "No MS2 spectra could be extracted for feature 3181 for the given mz and RT range"
#> [1] "Iteration 7"
#> [1] "No MS2 spectra could be extracted for feature 3819 for the given mz and RT range"
#> [1] "Iteration 8"
#> [1] "Iteration 9"
#> [1] "No MS2 spectra could be extracted for feature 1652 for the given mz and RT range"
#> [1] "Iteration 10"
#> [1] "No MS2 spectra could be extracted for feature 3677 for the given mz and RT range"
#> [1] "Iteration 11"
#> [1] "No MS2 spectra could be extracted for feature 2998 for the given mz and RT range"
#> [1] "Iteration 12"
#> [1] "No MS2 spectra could be extracted for feature 3726 for the given mz and RT range"
#> [1] "Iteration 13"
#> [1] "No MS2 spectra could be extracted for feature 3532 for the given mz and RT range"
#> [1] "Iteration 14"
#> [1] "No MS2 spectra could be extracted for feature 517 for the given mz and RT range"
#> [1] "Iteration 15"
#> [1] "No MS2 spectra could be extracted for feature 691 for the given mz and RT range"
#> [1] "Iteration 16"
#> [1] "No MS2 spectra could be extracted for feature 2790 for the given mz and RT range"
#> [1] "Iteration 17"
#> [1] "No MS2 spectra could be extracted for feature 4279 for the given mz and RT range"
#> [1] "Iteration 18"
#> [1] "No MS2 spectra could be extracted for feature 586 for the given mz and RT range"
#> [1] "Iteration 19"
#> [1] "No MS2 spectra could be extracted for feature 2364 for the given mz and RT range"
#> [1] "Iteration 20"
#> [1] "No MS2 spectra could be extracted for feature 4001 for the given mz and RT range"
#> [1] "Iteration 21"
#> [1] "No MS2 spectra could be extracted for feature 4116 for the given mz and RT range"
#> [1] "Iteration 22"
#> [1] "No MS2 spectra could be extracted for feature 1290 for the given mz and RT range"
#> [1] "Iteration 23"
#> [1] "No MS2 spectra could be extracted for feature 566 for the given mz and RT range"
#> [1] "Iteration 24"
#> [1] "Iteration 25"
#> [1] "No MS2 spectra could be extracted for feature 1319 for the given mz and RT range"
#> [1] "Iteration 26"
#> [1] "No MS2 spectra could be extracted for feature 4084 for the given mz and RT range"
#> [1] "Iteration 27"
#> [1] "No MS2 spectra could be extracted for feature 242 for the given mz and RT range"
#> [1] "Iteration 28"
#> [1] "No MS2 spectra could be extracted for feature 2385 for the given mz and RT range"
#> [1] "Iteration 29"
#> [1] "No MS2 spectra could be extracted for feature 782 for the given mz and RT range"
#> [1] "Iteration 30"
#> [1] "No MS2 spectra could be extracted for feature 534 for the given mz and RT range"
#> [1] "Iteration 31"
#> [1] "No MS2 spectra could be extracted for feature 1272 for the given mz and RT range"
#> [1] "Iteration 32"
#> [1] "No MS2 spectra could be extracted for feature 3930 for the given mz and RT range"
#> [1] "Iteration 33"
#> [1] "No MS2 spectra could be extracted for feature 4644 for the given mz and RT range"
#> [1] "Iteration 34"
#> [1] "Iteration 35"
#> [1] "No MS2 spectra could be extracted for feature 2815 for the given mz and RT range"
#> [1] "Iteration 36"
#> [1] "No MS2 spectra could be extracted for feature 2553 for the given mz and RT range"
#> [1] "Iteration 37"
#> [1] "No MS2 spectra could be extracted for feature 4589 for the given mz and RT range"
#> [1] "Iteration 38"
#> [1] "No MS2 spectra could be extracted for feature 4776 for the given mz and RT range"
#> [1] "Iteration 39"
#> [1] "No MS2 spectra could be extracted for feature 3462 for the given mz and RT range"
#> [1] "Iteration 40"
#> [1] "No MS2 spectra could be extracted for feature 57 for the given mz and RT range"
#> [1] "Iteration 41"
#> [1] "No MS2 spectra could be extracted for feature 806 for the given mz and RT range"
#> [1] "Iteration 42"
#> [1] "No MS2 spectra could be extracted for feature 3578 for the given mz and RT range"
#> [1] "Iteration 43"
#> [1] "No MS2 spectra could be extracted for feature 2597 for the given mz and RT range"
#> [1] "Iteration 44"
#> [1] "No MS2 spectra could be extracted for feature 4685 for the given mz and RT range"
#> [1] "Iteration 45"
#> [1] "No MS2 spectra could be extracted for feature 23 for the given mz and RT range"
#> [1] "Iteration 46"
#> [1] "No MS2 spectra could be extracted for feature 2263 for the given mz and RT range"
#> [1] "Iteration 47"
#> [1] "Iteration 48"
#> [1] "Iteration 49"
#> [1] "No MS2 spectra could be extracted for feature 1540 for the given mz and RT range"
#> [1] "Iteration 50"
#> [1] "No MS2 spectra could be extracted for feature 1142 for the given mz and RT range"
#> [1] "Iteration 51"
#> [1] "No MS2 spectra could be extracted for feature 226 for the given mz and RT range"
#> [1] "Iteration 52"
#> [1] "No MS2 spectra could be extracted for feature 708 for the given mz and RT range"
#> [1] "Iteration 53"
#> [1] "Iteration 54"
#> [1] "No MS2 spectra could be extracted for feature 414 for the given mz and RT range"
#> [1] "Iteration 55"
#> [1] "No MS2 spectra could be extracted for feature 316 for the given mz and RT range"
#> [1] "Iteration 56"
#> [1] "No MS2 spectra could be extracted for feature 1123 for the given mz and RT range"
#> [1] "Iteration 57"
#> [1] "No MS2 spectra could be extracted for feature 2983 for the given mz and RT range"
#> [1] "Iteration 58"
#> [1] "Iteration 59"
#> [1] "Iteration 60"
#> [1] "No MS2 spectra could be extracted for feature 3072 for the given mz and RT range"
#> [1] "Iteration 61"
#> [1] "Iteration 62"
#> [1] "No MS2 spectra could be extracted for feature 653 for the given mz and RT range"
#> [1] "Iteration 63"
#> [1] "Iteration 64"
#> [1] "No MS2 spectra could be extracted for feature 167 for the given mz and RT range"
#> [1] "Iteration 65"
#> [1] "No MS2 spectra could be extracted for feature 3687 for the given mz and RT range"
#> [1] "Iteration 66"
#> [1] "No MS2 spectra could be extracted for feature 2587 for the given mz and RT range"
#> [1] "Iteration 67"
#> [1] "No MS2 spectra could be extracted for feature 3859 for the given mz and RT range"
#> [1] "Iteration 68"
#> [1] "No MS2 spectra could be extracted for feature 1479 for the given mz and RT range"
#> [1] "Iteration 69"
#> [1] "No MS2 spectra could be extracted for feature 3913 for the given mz and RT range"
#> [1] "Iteration 70"
#> [1] "Iteration 71"
#> [1] "No MS2 spectra could be extracted for feature 1634 for the given mz and RT range"
#> [1] "Iteration 72"
#> [1] "No MS2 spectra could be extracted for feature 2220 for the given mz and RT range"
#> [1] "Iteration 73"
#> [1] "No MS2 spectra could be extracted for feature 1381 for the given mz and RT range"
#> [1] "Iteration 74"
#> [1] "No MS2 spectra could be extracted for feature 3447 for the given mz and RT range"
#> [1] "Iteration 75"
#> [1] "Iteration 76"
#> [1] "No MS2 spectra could be extracted for feature 655 for the given mz and RT range"
#> [1] "Iteration 77"
#> [1] "Iteration 78"
#> [1] "No MS2 spectra could be extracted for feature 3731 for the given mz and RT range"
#> [1] "Iteration 79"
#> [1] "No MS2 spectra could be extracted for feature 3224 for the given mz and RT range"
#> [1] "Iteration 80"
#> [1] "Iteration 81"
#> [1] "No MS2 spectra could be extracted for feature 2234 for the given mz and RT range"
#> [1] "Iteration 82"
#> [1] "Iteration 83"
#> [1] "No MS2 spectra could be extracted for feature 3292 for the given mz and RT range"
#> [1] "Iteration 84"
#> [1] "No MS2 spectra could be extracted for feature 691 for the given mz and RT range"
#> [1] "Iteration 85"
#> [1] "Iteration 86"
#> [1] "No MS2 spectra could be extracted for feature 425 for the given mz and RT range"
#> [1] "Iteration 87"
#> [1] "No MS2 spectra could be extracted for feature 841 for the given mz and RT range"
#> [1] "Iteration 88"
#> [1] "No MS2 spectra could be extracted for feature 486 for the given mz and RT range"
#> [1] "Iteration 89"
#> [1] "No MS2 spectra could be extracted for feature 2124 for the given mz and RT range"
#> [1] "Iteration 90"
#> [1] "No MS2 spectra could be extracted for feature 3440 for the given mz and RT range"
#> [1] "Iteration 91"
#> [1] "No MS2 spectra could be extracted for feature 3830 for the given mz and RT range"
#> [1] "Iteration 92"
#> [1] "No MS2 spectra could be extracted for feature 3766 for the given mz and RT range"
#> [1] "Iteration 93"
#> [1] "No MS2 spectra could be extracted for feature 823 for the given mz and RT range"
#> [1] "Iteration 94"
#> [1] "Iteration 95"
#> [1] "No MS2 spectra could be extracted for feature 3536 for the given mz and RT range"
#> [1] "Iteration 96"
#> [1] "No MS2 spectra could be extracted for feature 1856 for the given mz and RT range"
#> [1] "Iteration 97"
#> [1] "No MS2 spectra could be extracted for feature 444 for the given mz and RT range"
#> [1] "Iteration 98"
#> [1] "No MS2 spectra could be extracted for feature 944 for the given mz and RT range"
#> [1] "Iteration 99"
#> [1] "No MS2 spectra could be extracted for feature 1040 for the given mz and RT range"
#> [1] "Iteration 100"
#> [1] "No MS2 spectra could be extracted for feature 1244 for the given mz and RT range"
#> [1] "MS2 spectra couldn't be extracted from a total of82 features out of 100 features. They are as follows:"
#> 2502,4552,3794,3941,3181,3819,1652,3677,2998,3726,3532,517,691,2790,4279,586,2364,4001,4116,1290,566,1319,4084,242,2385,782,534,1272,3930,4644,2815,2553,4589,4776,3462,57,806,3578,2597,4685,23,2263,1540,1142,226,708,414,316,1123,2983,3072,653,167,3687,2587,3859,1479,3913,1634,2220,1381,3447,655,3731,3224,2234,3292,691,425,841,486,2124,3440,3830,3766,823,3536,1856,444,944,1040,1244
l <- sp[[1]]
sp1 <- extract_raw_spectra(folder_path, l, 0.05, 0.8)
#> Error in txtProgressBar(min = 0, max = length(sps_aggregate_all_mets),     style = 3): must have 'max' > 'min'
spectra_list <- sp1[[1]]
#> Error in eval(expr, envir, enclos): object 'sp1' not found
call_aggregate(spectra_list, 0.05, folder_path)
#> Error in eval(expr, envir, enclos): object 'spectra_list' not found