call_aggregate

Helper code to run derive_aggregate_spectra

call_aggregate(spectra_list, mz_tol, folder_path)

Arguments

spectra_list:

list of spectra objects for all features, each containing top x% TIC spectra concatenated

mz_tol:

mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature

folder_path:

the path to the folder where you want to store the data reduction dataframe

Value

a list containing the aggregate spectra object and a dataframe of the aggregate spectra for all features

Examples

# Example usage of the function
folder_path = "/home/shayantan/Desktop/test_package_dures/test_2/"
sp <- preprocess(folder_path, 5, 0.01)
#> *******************Sanity check for input files***********************
#> ✓ Stats file found!
#> ✓ Stats file has the correct columns
#> ✓ mzml folder exists
#> RT tolerance file not found. Proceeding with the default mz tolerance of: 5 ppm and RT tolerance of: 0.01 minutes for all samples
#> Reading mzml files
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
#>  MS2 data are not present in the following files:
#> [1] "Blank_01,Blank_02,x200708x_Pat_01_POS,x200708x_Pat_02_POS,x200708x_Pat_03_POS,x200708x_Pat_04_POS,x200708x_Pat_05_POS,x200708x_Pat_06_POS,x200708x_Pat_07_POS,x200708x_Pat_08_POS,x200708x_Pat_09_POS,x200708x_Pat_100_POS,x200708x_Pat_101_POS,x200708x_Pat_102_POS,x200708x_Pat_103_POS,x200708x_Pat_104_POS,x200708x_Pat_105_POS,x200708x_Pat_106_POS,x200708x_Pat_107_POS,x200708x_Pat_108_POS,x200708x_Pat_109_POS,x200708x_Pat_10_POS,x200708x_Pat_110_POS,x200708x_Pat_111_POS,x200708x_Pat_112_POS,x200708x_Pat_114_POS,x200708x_Pat_115_POS,x200708x_Pat_116_POS,x200708x_Pat_117_POS,x200708x_Pat_118_POS,x200708x_Pat_119_POS,x200708x_Pat_11_POS,x200708x_Pat_120_POS,x200708x_Pat_121_POS,x200708x_Pat_122_POS,x200708x_Pat_123_POS,x200708x_Pat_124_POS,x200708x_Pat_125_POS,x200708x_Pat_126_POS,x200708x_Pat_127_POS,x200708x_Pat_128_POS,x200708x_Pat_129_POS,x200708x_Pat_12_POS,x200708x_Pat_130_POS,x200708x_Pat_131_POS,x200708x_Pat_132_POS,x200708x_Pat_133_POS,x200708x_Pat_135_POS,x200708x_Pat_137_POS,x200708x_Pat_138_POS,x200708x_Pat_139_POS,x200708x_Pat_13_POS,x200708x_Pat_140_POS,x200708x_Pat_14_POS,x200708x_Pat_15_POS,x200708x_Pat_16_POS,x200708x_Pat_18_POS,x200708x_Pat_19_POS,x200708x_Pat_20_POS,x200708x_Pat_22_POS,x200708x_Pat_23_POS,x200708x_Pat_26_POS,x200708x_Pat_27_POS,x200708x_Pat_28_POS,x200708x_Pat_32_POS,x200708x_Pat_33_POS,x200708x_Pat_34_POS,x200708x_Pat_35_POS,x200708x_Pat_36_POS,x200708x_Pat_37_POS,x200708x_Pat_38_POS,x200708x_Pat_39_POS,x200708x_Pat_40_POS,x200708x_Pat_41_POS,x200708x_Pat_42_POS,x200708x_Pat_43_POS,x200708x_Pat_44_POS,x200708x_Pat_45_POS,x200708x_Pat_46_POS,x200708x_Pat_47_POS,x200708x_Pat_48_POS,x200708x_Pat_49_POS,x200708x_Pat_50_POS,x200708x_Pat_51_POS,x200708x_Pat_52_POS,x200708x_Pat_53_POS,x200708x_Pat_54_POS,x200708x_Pat_55_POS,x200708x_Pat_56_POS,x200708x_Pat_57_POS,x200708x_Pat_58_POS,x200708x_Pat_59_POS,x200708x_Pat_60_POS,x200708x_Pat_61_POS,x200708x_Pat_62_POS,x200708x_Pat_63_POS,x200708x_Pat_64_POS,x200708x_Pat_65_POS,x200708x_Pat_66_POS,x200708x_Pat_67_POS,x200708x_Pat_68_POS,x200708x_Pat_69_POS,x200708x_Pat_70_POS,x200708x_Pat_71_POS,x200708x_Pat_72_POS,x200708x_Pat_73_POS,x200708x_Pat_74_POS,x200708x_Pat_75_POS,x200708x_Pat_76_POS,x200708x_Pat_77_POS,x200708x_Pat_78_POS,x200708x_Pat_79_POS,x200708x_Pat_80_POS,x200708x_Pat_81_POS,x200708x_Pat_82_POS,x200708x_Pat_83_POS,x200708x_Pat_84_POS,x200708x_Pat_85_POS,x200708x_Pat_86_POS,x200708x_Pat_87_POS,x200708x_Pat_89_POS,x200708x_Pat_90_POS,x200708x_Pat_91_POS,x200708x_Pat_92_POS,x200708x_Pat_93_POS,x200708x_Pat_94_POS,x200708x_Pat_95_POS,x200708x_Pat_96_POS,x200708x_Pat_97_POS,x200708x_Pat_98_POS,x200708x_Pat_99_POS,x200708x_QC10_POS,x200708x_QC11_POS,x200708x_QC12_POS,x200708x_QC13_POS,x200708x_QC14_POS,x200708x_QC15_POS,x200708x_QC1_POS,x200708x_QC2_POS,x200708x_QC3_POS,x200708x_QC4_POS,x200708x_QC5_POS,x200708x_QC6_POS,x200708x_QC7_POS,x200708x_QC8_POS,x200708x_QC9_POS"
#> ✓ Spectra has been extracted from the mzml files[1] "Iteration 1"
#> [1] "No MS2 spectra could be extracted for feature 2502 for the given mz and RT range"
#> [1] "Iteration 2"
#> [1] "Iteration 3"
#> [1] "No MS2 spectra could be extracted for feature 4552 for the given mz and RT range"
#> [1] "Iteration 4"
#> [1] "No MS2 spectra could be extracted for feature 3794 for the given mz and RT range"
#> [1] "Iteration 5"
#> [1] "No MS2 spectra could be extracted for feature 3941 for the given mz and RT range"
#> [1] "Iteration 6"
#> [1] "No MS2 spectra could be extracted for feature 3181 for the given mz and RT range"
#> [1] "Iteration 7"
#> [1] "No MS2 spectra could be extracted for feature 3819 for the given mz and RT range"
#> [1] "Iteration 8"
#> [1] "Iteration 9"
#> [1] "No MS2 spectra could be extracted for feature 1652 for the given mz and RT range"
#> [1] "Iteration 10"
#> [1] "No MS2 spectra could be extracted for feature 3677 for the given mz and RT range"
#> [1] "Iteration 11"
#> [1] "No MS2 spectra could be extracted for feature 2998 for the given mz and RT range"
#> [1] "Iteration 12"
#> [1] "No MS2 spectra could be extracted for feature 3726 for the given mz and RT range"
#> [1] "Iteration 13"
#> [1] "No MS2 spectra could be extracted for feature 3532 for the given mz and RT range"
#> [1] "Iteration 14"
#> [1] "No MS2 spectra could be extracted for feature 517 for the given mz and RT range"
#> [1] "Iteration 15"
#> [1] "No MS2 spectra could be extracted for feature 691 for the given mz and RT range"
#> [1] "Iteration 16"
#> [1] "No MS2 spectra could be extracted for feature 2790 for the given mz and RT range"
#> [1] "Iteration 17"
#> [1] "No MS2 spectra could be extracted for feature 4279 for the given mz and RT range"
#> [1] "Iteration 18"
#> [1] "No MS2 spectra could be extracted for feature 586 for the given mz and RT range"
#> [1] "Iteration 19"
#> [1] "No MS2 spectra could be extracted for feature 2364 for the given mz and RT range"
#> [1] "Iteration 20"
#> [1] "No MS2 spectra could be extracted for feature 4001 for the given mz and RT range"
#> [1] "Iteration 21"
#> [1] "No MS2 spectra could be extracted for feature 4116 for the given mz and RT range"
#> [1] "Iteration 22"
#> [1] "No MS2 spectra could be extracted for feature 1290 for the given mz and RT range"
#> [1] "Iteration 23"
#> [1] "No MS2 spectra could be extracted for feature 566 for the given mz and RT range"
#> [1] "Iteration 24"
#> [1] "Iteration 25"
#> [1] "No MS2 spectra could be extracted for feature 1319 for the given mz and RT range"
#> [1] "Iteration 26"
#> [1] "No MS2 spectra could be extracted for feature 4084 for the given mz and RT range"
#> [1] "Iteration 27"
#> [1] "No MS2 spectra could be extracted for feature 242 for the given mz and RT range"
#> [1] "Iteration 28"
#> [1] "No MS2 spectra could be extracted for feature 2385 for the given mz and RT range"
#> [1] "Iteration 29"
#> [1] "No MS2 spectra could be extracted for feature 782 for the given mz and RT range"
#> [1] "Iteration 30"
#> [1] "No MS2 spectra could be extracted for feature 534 for the given mz and RT range"
#> [1] "Iteration 31"
#> [1] "No MS2 spectra could be extracted for feature 1272 for the given mz and RT range"
#> [1] "Iteration 32"
#> [1] "No MS2 spectra could be extracted for feature 3930 for the given mz and RT range"
#> [1] "Iteration 33"
#> [1] "No MS2 spectra could be extracted for feature 4644 for the given mz and RT range"
#> [1] "Iteration 34"
#> [1] "Iteration 35"
#> [1] "No MS2 spectra could be extracted for feature 2815 for the given mz and RT range"
#> [1] "Iteration 36"
#> [1] "No MS2 spectra could be extracted for feature 2553 for the given mz and RT range"
#> [1] "Iteration 37"
#> [1] "No MS2 spectra could be extracted for feature 4589 for the given mz and RT range"
#> [1] "Iteration 38"
#> [1] "No MS2 spectra could be extracted for feature 4776 for the given mz and RT range"
#> [1] "Iteration 39"
#> [1] "No MS2 spectra could be extracted for feature 3462 for the given mz and RT range"
#> [1] "Iteration 40"
#> [1] "No MS2 spectra could be extracted for feature 57 for the given mz and RT range"
#> [1] "Iteration 41"
#> [1] "No MS2 spectra could be extracted for feature 806 for the given mz and RT range"
#> [1] "Iteration 42"
#> [1] "No MS2 spectra could be extracted for feature 3578 for the given mz and RT range"
#> [1] "Iteration 43"
#> [1] "No MS2 spectra could be extracted for feature 2597 for the given mz and RT range"
#> [1] "Iteration 44"
#> [1] "No MS2 spectra could be extracted for feature 4685 for the given mz and RT range"
#> [1] "Iteration 45"
#> [1] "No MS2 spectra could be extracted for feature 23 for the given mz and RT range"
#> [1] "Iteration 46"
#> [1] "No MS2 spectra could be extracted for feature 2263 for the given mz and RT range"
#> [1] "Iteration 47"
#> [1] "Iteration 48"
#> [1] "Iteration 49"
#> [1] "No MS2 spectra could be extracted for feature 1540 for the given mz and RT range"
#> [1] "Iteration 50"
#> [1] "No MS2 spectra could be extracted for feature 1142 for the given mz and RT range"
#> [1] "Iteration 51"
#> [1] "No MS2 spectra could be extracted for feature 226 for the given mz and RT range"
#> [1] "Iteration 52"
#> [1] "No MS2 spectra could be extracted for feature 708 for the given mz and RT range"
#> [1] "Iteration 53"
#> [1] "Iteration 54"
#> [1] "No MS2 spectra could be extracted for feature 414 for the given mz and RT range"
#> [1] "Iteration 55"
#> [1] "No MS2 spectra could be extracted for feature 316 for the given mz and RT range"
#> [1] "Iteration 56"
#> [1] "No MS2 spectra could be extracted for feature 1123 for the given mz and RT range"
#> [1] "Iteration 57"
#> [1] "No MS2 spectra could be extracted for feature 2983 for the given mz and RT range"
#> [1] "Iteration 58"
#> [1] "Iteration 59"
#> [1] "Iteration 60"
#> [1] "No MS2 spectra could be extracted for feature 3072 for the given mz and RT range"
#> [1] "Iteration 61"
#> [1] "Iteration 62"
#> [1] "No MS2 spectra could be extracted for feature 653 for the given mz and RT range"
#> [1] "Iteration 63"
#> [1] "Iteration 64"
#> [1] "No MS2 spectra could be extracted for feature 167 for the given mz and RT range"
#> [1] "Iteration 65"
#> [1] "No MS2 spectra could be extracted for feature 3687 for the given mz and RT range"
#> [1] "Iteration 66"
#> [1] "No MS2 spectra could be extracted for feature 2587 for the given mz and RT range"
#> [1] "Iteration 67"
#> [1] "No MS2 spectra could be extracted for feature 3859 for the given mz and RT range"
#> [1] "Iteration 68"
#> [1] "No MS2 spectra could be extracted for feature 1479 for the given mz and RT range"
#> [1] "Iteration 69"
#> [1] "No MS2 spectra could be extracted for feature 3913 for the given mz and RT range"
#> [1] "Iteration 70"
#> [1] "Iteration 71"
#> [1] "No MS2 spectra could be extracted for feature 1634 for the given mz and RT range"
#> [1] "Iteration 72"
#> [1] "No MS2 spectra could be extracted for feature 2220 for the given mz and RT range"
#> [1] "Iteration 73"
#> [1] "No MS2 spectra could be extracted for feature 1381 for the given mz and RT range"
#> [1] "Iteration 74"
#> [1] "No MS2 spectra could be extracted for feature 3447 for the given mz and RT range"
#> [1] "Iteration 75"
#> [1] "Iteration 76"
#> [1] "No MS2 spectra could be extracted for feature 655 for the given mz and RT range"
#> [1] "Iteration 77"
#> [1] "Iteration 78"
#> [1] "No MS2 spectra could be extracted for feature 3731 for the given mz and RT range"
#> [1] "Iteration 79"
#> [1] "No MS2 spectra could be extracted for feature 3224 for the given mz and RT range"
#> [1] "Iteration 80"
#> [1] "Iteration 81"
#> [1] "No MS2 spectra could be extracted for feature 2234 for the given mz and RT range"
#> [1] "Iteration 82"
#> [1] "Iteration 83"
#> [1] "No MS2 spectra could be extracted for feature 3292 for the given mz and RT range"
#> [1] "Iteration 84"
#> [1] "No MS2 spectra could be extracted for feature 691 for the given mz and RT range"
#> [1] "Iteration 85"
#> [1] "Iteration 86"
#> [1] "No MS2 spectra could be extracted for feature 425 for the given mz and RT range"
#> [1] "Iteration 87"
#> [1] "No MS2 spectra could be extracted for feature 841 for the given mz and RT range"
#> [1] "Iteration 88"
#> [1] "No MS2 spectra could be extracted for feature 486 for the given mz and RT range"
#> [1] "Iteration 89"
#> [1] "No MS2 spectra could be extracted for feature 2124 for the given mz and RT range"
#> [1] "Iteration 90"
#> [1] "No MS2 spectra could be extracted for feature 3440 for the given mz and RT range"
#> [1] "Iteration 91"
#> [1] "No MS2 spectra could be extracted for feature 3830 for the given mz and RT range"
#> [1] "Iteration 92"
#> [1] "No MS2 spectra could be extracted for feature 3766 for the given mz and RT range"
#> [1] "Iteration 93"
#> [1] "No MS2 spectra could be extracted for feature 823 for the given mz and RT range"
#> [1] "Iteration 94"
#> [1] "Iteration 95"
#> [1] "No MS2 spectra could be extracted for feature 3536 for the given mz and RT range"
#> [1] "Iteration 96"
#> [1] "No MS2 spectra could be extracted for feature 1856 for the given mz and RT range"
#> [1] "Iteration 97"
#> [1] "No MS2 spectra could be extracted for feature 444 for the given mz and RT range"
#> [1] "Iteration 98"
#> [1] "No MS2 spectra could be extracted for feature 944 for the given mz and RT range"
#> [1] "Iteration 99"
#> [1] "No MS2 spectra could be extracted for feature 1040 for the given mz and RT range"
#> [1] "Iteration 100"
#> [1] "No MS2 spectra could be extracted for feature 1244 for the given mz and RT range"
#> [1] "MS2 spectra couldn't be extracted from a total of 82 features out of 100 features. They are as follows:"
#> 2502
#> 4552
#> 3794
#> 3941
#> 3181
#> 3819
#> 1652
#> 3677
#> 2998
#> 3726
#> 3532
#> 517
#> 691
#> 2790
#> 4279
#> 586
#> 2364
#> 4001
#> 4116
#> 1290
#> 566
#> 1319
#> 4084
#> 242
#> 2385
#> 782
#> 534
#> 1272
#> 3930
#> 4644
#> 2815
#> 2553
#> 4589
#> 4776
#> 3462
#> 57
#> 806
#> 3578
#> 2597
#> 4685
#> 23
#> 2263
#> 1540
#> 1142
#> 226
#> 708
#> 414
#> 316
#> 1123
#> 2983
#> 3072
#> 653
#> 167
#> 3687
#> 2587
#> 3859
#> 1479
#> 3913
#> 1634
#> 2220
#> 1381
#> 3447
#> 655
#> 3731
#> 3224
#> 2234
#> 3292
#> 691
#> 425
#> 841
#> 486
#> 2124
#> 3440
#> 3830
#> 3766
#> 823
#> 3536
#> 1856
#> 444
#> 944
#> 1040
#> 1244
l <- sp[[1]]
sp1 <- extract_raw_spectra(folder_path, l, 0.05, 0.8)
#> Error in txtProgressBar(min = 0, max = length(sps_aggregate_all_mets),     style = 3): must have 'max' > 'min'
spectra_list <- sp1[[1]]
#> Error in eval(expr, envir, enclos): object 'sp1' not found
call_aggregate(spectra_list, 0.05, folder_path)
#> Error in eval(expr, envir, enclos): object 'spectra_list' not found