call_aggregate

Helper code to run derive_aggregate_spectra

call_aggregate(spectra_list, mz_tol, folder_path)

Arguments

spectra_list:

list of spectra objects for all features, each containing top x% TIC spectra concatenated

mz_tol:

mass tolerance (in Da) required when grouping fragments across multiple spectra for a given feature

folder_path:

the path to the folder where you want to store the data reduction dataframe

Value

a list containing the aggregate spectra object and a dataframe of the aggregate spectra for all features

Examples

# Example usage of the function
folder_path = "/home/shayantan/Desktop/test_package_dures/test_2/"
sp <- preprocess(folder_path, 5, 0.01)
#> *******************Sanity check for input files***********************
#> ✓ Feature list file found!
#> ✓ Feature list file has the correct columns
#> ✓ mzml folder exists
#> RT tolerance file not found. Proceeding with the default mz tolerance of: 5 ppm and RT tolerance of: 0.01 minutes for all samples
#> Reading mzml files
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
#>  MS2 data are not present in the following files:
#> [1] "Blank_01,Blank_02,x200708x_Pat_01_POS,x200708x_Pat_02_POS,x200708x_Pat_03_POS,x200708x_Pat_04_POS,x200708x_Pat_05_POS,x200708x_Pat_06_POS,x200708x_Pat_07_POS,x200708x_Pat_08_POS,x200708x_Pat_09_POS,x200708x_Pat_100_POS,x200708x_Pat_101_POS,x200708x_Pat_102_POS,x200708x_Pat_103_POS,x200708x_Pat_104_POS,x200708x_Pat_105_POS,x200708x_Pat_106_POS,x200708x_Pat_107_POS,x200708x_Pat_108_POS,x200708x_Pat_109_POS,x200708x_Pat_10_POS,x200708x_Pat_110_POS,x200708x_Pat_111_POS,x200708x_Pat_112_POS,x200708x_Pat_114_POS,x200708x_Pat_115_POS,x200708x_Pat_116_POS,x200708x_Pat_117_POS,x200708x_Pat_118_POS,x200708x_Pat_119_POS,x200708x_Pat_11_POS,x200708x_Pat_120_POS,x200708x_Pat_121_POS,x200708x_Pat_122_POS,x200708x_Pat_123_POS,x200708x_Pat_124_POS,x200708x_Pat_125_POS,x200708x_Pat_126_POS,x200708x_Pat_127_POS,x200708x_Pat_128_POS,x200708x_Pat_129_POS,x200708x_Pat_12_POS,x200708x_Pat_130_POS,x200708x_Pat_131_POS,x200708x_Pat_132_POS,x200708x_Pat_133_POS,x200708x_Pat_135_POS,x200708x_Pat_137_POS,x200708x_Pat_138_POS,x200708x_Pat_139_POS,x200708x_Pat_13_POS,x200708x_Pat_140_POS,x200708x_Pat_14_POS,x200708x_Pat_15_POS,x200708x_Pat_16_POS,x200708x_Pat_18_POS,x200708x_Pat_19_POS,x200708x_Pat_20_POS,x200708x_Pat_22_POS,x200708x_Pat_23_POS,x200708x_Pat_26_POS,x200708x_Pat_27_POS,x200708x_Pat_28_POS,x200708x_Pat_32_POS,x200708x_Pat_33_POS,x200708x_Pat_34_POS,x200708x_Pat_35_POS,x200708x_Pat_36_POS,x200708x_Pat_37_POS,x200708x_Pat_38_POS,x200708x_Pat_39_POS,x200708x_Pat_40_POS,x200708x_Pat_41_POS,x200708x_Pat_42_POS,x200708x_Pat_43_POS,x200708x_Pat_44_POS,x200708x_Pat_45_POS,x200708x_Pat_46_POS,x200708x_Pat_47_POS,x200708x_Pat_48_POS,x200708x_Pat_49_POS,x200708x_Pat_50_POS,x200708x_Pat_51_POS,x200708x_Pat_52_POS,x200708x_Pat_53_POS,x200708x_Pat_54_POS,x200708x_Pat_55_POS,x200708x_Pat_56_POS,x200708x_Pat_57_POS,x200708x_Pat_58_POS,x200708x_Pat_59_POS,x200708x_Pat_60_POS,x200708x_Pat_61_POS,x200708x_Pat_62_POS,x200708x_Pat_63_POS,x200708x_Pat_64_POS,x200708x_Pat_65_POS,x200708x_Pat_66_POS,x200708x_Pat_67_POS,x200708x_Pat_68_POS,x200708x_Pat_69_POS,x200708x_Pat_70_POS,x200708x_Pat_71_POS,x200708x_Pat_72_POS,x200708x_Pat_73_POS,x200708x_Pat_74_POS,x200708x_Pat_75_POS,x200708x_Pat_76_POS,x200708x_Pat_77_POS,x200708x_Pat_78_POS,x200708x_Pat_79_POS,x200708x_Pat_80_POS,x200708x_Pat_81_POS,x200708x_Pat_82_POS,x200708x_Pat_83_POS,x200708x_Pat_84_POS,x200708x_Pat_85_POS,x200708x_Pat_86_POS,x200708x_Pat_87_POS,x200708x_Pat_89_POS,x200708x_Pat_90_POS,x200708x_Pat_91_POS,x200708x_Pat_92_POS,x200708x_Pat_93_POS,x200708x_Pat_94_POS,x200708x_Pat_95_POS,x200708x_Pat_96_POS,x200708x_Pat_97_POS,x200708x_Pat_98_POS,x200708x_Pat_99_POS,x200708x_QC10_POS,x200708x_QC11_POS,x200708x_QC12_POS,x200708x_QC13_POS,x200708x_QC14_POS,x200708x_QC15_POS,x200708x_QC1_POS,x200708x_QC2_POS,x200708x_QC3_POS,x200708x_QC4_POS,x200708x_QC5_POS,x200708x_QC6_POS,x200708x_QC7_POS,x200708x_QC8_POS,x200708x_QC9_POS"
#> 
#>  ✓ Spectra has been extracted from the mzml files
#> Concatenating and checking if MS2 data is available for a given feature
#> MS2 spectra couldn't be extracted from a total of 82 features out of 100 features. They are as follows:2502,4552,3794,3941,3181,3819,1652,3677,2998,3726,3532,517,691,2790,4279,586,2364,4001,4116,1290,566,1319,4084,242,2385,782,534,1272,3930,4644,2815,2553,4589,4776,3462,57,806,3578,2597,4685,23,2263,1540,1142,226,708,414,316,1123,2983,3072,653,167,3687,2587,3859,1479,3913,1634,2220,1381,3447,655,3731,3224,2234,3292,691,425,841,486,2124,3440,3830,3766,823,3536,1856,444,944,1040,1244
l <- sp[[1]]
sp1 <- extract_raw_spectra(folder_path, l, 0.05, 0.8)
#> Error in txtProgressBar(min = 0, max = length(sps_aggregate_all_mets),     style = 3): must have 'max' > 'min'
spectra_list <- sp1[[1]]
#> Error in eval(expr, envir, enclos): object 'sp1' not found
call_aggregate(spectra_list, 0.05, folder_path)
#> Error in eval(expr, envir, enclos): object 'spectra_list' not found